Fixing file split
